kitty: install components to build kitty from source
kitty requires a few X11 libraries, as well as a golang SDK. Fixed my
install to build /usr/local/go directory. Along the way learned that
.zprofile is only read for login shells, not interactive terminal
emulator sessions - so I moved the content from my .zprofile to
.zshenv.

At the end of this, I discovered kitty has some weird build
contraints (pythin dep?!?!), and the "simple" build/execute isn't so
simple - so i gave up trying to build it myself.

Might need to just update to arch for a version of kitty that is not >
4years old (thanks, pop!_os being stuck on ubuntu 22.04)